refactor(lang): Move CommandGenerator and its friends to a proper home
